feat: update Hero component styles and refactor useInView hook import; add throttle utility
