+++ components/hooks/useInView.ts
@@ -1,0 +1,2 @@
+export { default } from "@/hooks/useInView";
+
